Anfreq function, constants cleanup, thermochemcalc finished, ORCAfreqgrab, geometric-1atom option,

+++ interface_geometric.py
@@ -23,7 +23,16 @@ def geomeTRICOptimizer(theory=None,fragment=None, coordsystem='tric', frozenatom
     Active-atom coords (e.g. only QM region) are only provided to geomeTRIC during optimization while rest is frozen.
     Needed as discussed here: https://github.com/leeping/geomeTRIC/commit/584869707aca1dbeabab6fe873fdf139d384ca66#diff-2af7dd72b77dac63cea64c052a549fe0
     """
-    
+
+    if fragment.numatoms == 1:
+        print("System has 1 atoms.")
+        print("Doing single-point energy calculation instead")
+        energy = Singlepoint(fragment=fragment, theory=theory)
+        return energy
+        #E = self.theory.run(current_coords=fragment.coords, elems=fragment.elems, Grad=False)
+
+
+
     try:
         os.remove('geometric_OPTtraj.log')
         os.remove('geometric_OPTtraj.xyz')
@@ -35,7 +44,6 @@ def geomeTRICOptimizer(theory=None,fragment=None, coordsystem='tric', frozenatom
     except:
         pass
     
-    # TODO:
 
     #NOTE: We are now sorting actatoms and qmatoms list both here and in QM/MM object
     #: Alternatively we could sort the actatoms list and qmatoms list in QM/MM object before doing anything. Need to check carefully though....
@@ -96,16 +104,16 @@ def geomeTRICOptimizer(theory=None,fragment=None, coordsystem='tric', frozenatom
         print("Number of active atoms:", len(actatoms))
         actcoords, actelems = fragment.get_coords_for_atoms(actatoms)
         
-        #Writing act-region coords (only) of Yggdrasill fragment to disk as XYZ file and reading into geomeTRIC
+        #Writing act-region coords (only) of ASH fragment to disk as XYZ file and reading into geomeTRIC
         write_xyzfile(actelems, actcoords, 'initialxyzfiletric')
         mol_geometric_frag=geometric.molecule.Molecule("initialxyzfiletric.xyz")
     else:
-        #Write coordinates from Yggdrasill fragment to disk as XYZ-file and reading into geomeTRIC
+        #Write coordinates from ASH fragment to disk as XYZ-file and reading into geomeTRIC
         fragment.write_xyzfile("initialxyzfiletric.xyz")
         mol_geometric_frag=geometric.molecule.Molecule("initialxyzfiletric.xyz")
 
-    #Defining Yggdrasill engine class used to communicate with geomeTRIC
-    class Yggdrasillengineclass:
+    #Defining ASH engine class used to communicate with geomeTRIC
+    class ASHengineclass:
         def __init__(self,geometric_molf, theory, ActiveRegion=False, actatoms=None):
             #Defining M attribute of engine object as geomeTRIC Molecule object
             self.M=geometric_molf
@@ -212,7 +220,7 @@ def geomeTRICOptimizer(theory=None,fragment=None, coordsystem='tric', frozenatom
         with open("constraints.txt", 'a') as confile:
             confile.write('$freeze\n')
             for frozat in frozenatoms:
-                #Changing from zero-indexing (Yggdrasill) to 1-indexing (geomeTRIC)
+                #Changing from zero-indexing (ASH) to 1-indexing (geomeTRIC)
                 frozenatomindex=frozat+1
                 confile.write('xyz {}\n'.format(frozenatomindex))
     #Bond constraints
@@ -296,8 +304,8 @@ def geomeTRICOptimizer(theory=None,fragment=None, coordsystem='tric', frozenatom
 
 
 
-    #Defining Yggdrasill engine object containing geometry and theory. ActiveRegion boolean passed.
-    ashengine = Yggdrasillengineclass(mol_geometric_frag,theory, ActiveRegion=ActiveRegion, actatoms=actatoms)
+    #Defining ASHengineclass engine object containing geometry and theory. ActiveRegion boolean passed.
+    ashengine = ASHengineclass(mol_geometric_frag,theory, ActiveRegion=ActiveRegion, actatoms=actatoms)
     #Defining args object, containing engine object
     args=geomeTRICArgsObject(ashengine,constraintsfile,coordsys=coordsystem, maxiter=maxiter, conv_criteria=conv_criteria)
